+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Added Sentry Error reporting to API
 
+- Added more information to UI Sentry
+
 ## [v0.8.1] - 2018-11-14
 
 * Changed: InstanceInfo now should be passed in Url-Encoding